Move validation to own file
