+++ DocumentFormat.OpenXml/Validation/Schema/AlternateContentValidator.cs
@@ -2,8 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using DocumentFormat.OpenXml.Validation;
-using System.Collections.Generic;
-using System.Diagnostics;
 
 namespace DocumentFormat.OpenXml.Internal.SchemaValidation
 {
@@ -189,302 +187,5 @@ namespace DocumentFormat.OpenXml.Internal.SchemaValidation
             }
             return false;
         }
-    }
-
-    /// <summary>
-    /// Compatibility-Rule Attributes
-    /// </summary>
-    internal static class CompatibilityRuleAttributesValidator
-    {
-        /// <summary>
-        /// Validate compatibility rule attributes - Ignorable, ProcessContent, PreserveElements, PreserveAttributes, MustUnderstand.
-        /// </summary>
-        /// <param name="validationContext">The validation context.</param>
-        internal static void ValidateMcAttributes(ValidationContext validationContext)
-        {
-            var element = validationContext.Element;
-            if (element.MCAttributes == null)
-            {
-                return;
-            }
-
-            HashSet<string> ignorableNamespaces = null;
-            ValidationErrorInfo errorInfo;
-
-            if (element.MCAttributes != null)
-            {
-                // validate Ignorable attribute
-                if (!string.IsNullOrEmpty(element.MCAttributes.Ignorable))
-                {
-                    ignorableNamespaces = new HashSet<string>();
-
-                    // rule: the prefix must already be defined.
-                    var prefixes = new ListValue<StringValue>();
-                    prefixes.InnerText = element.MCAttributes.Ignorable;
-                    foreach (var prefix in prefixes.Items)
-                    {
-                        var ignorableNamespace = element.LookupNamespace(prefix);
-                        if (string.IsNullOrEmpty(ignorableNamespace))
-                        {
-                            // error, the prefix is not defined.
-                            errorInfo = validationContext.ComposeMcValidationError(element, "MC_InvalidIgnorableAttribute", element.MCAttributes.Ignorable);
-                            validationContext.EmitError(errorInfo);
-                        }
-                        else
-                        {
-                            ignorableNamespaces.Add(ignorableNamespace);
-                        }
-                    }
-                }
-
-                // validate PreserveAttributes attribute
-                if (!string.IsNullOrEmpty(element.MCAttributes.PreserveAttributes))
-                {
-                    // The ProcessAttributes attribute value shall not reference any attribute name that does not belong to a namespace
-                    // that is identified by the Ignorable attribute of the same element.
-                    if (ignorableNamespaces == null)
-                    {
-                        // must have Ignorable on same element.
-                        errorInfo = validationContext.ComposeMcValidationError(element, "MC_InvalidPreserveAttributesAttribute", element.MCAttributes.PreserveAttributes);
-                        validationContext.EmitError(errorInfo);
-                    }
-                    else
-                    {
-                        string errorQName = ValidateQNameList(element.MCAttributes.PreserveAttributes, ignorableNamespaces, validationContext);
-                        if (!string.IsNullOrEmpty(errorQName))
-                        {
-                            errorInfo = validationContext.ComposeMcValidationError(element, "MC_InvalidPreserveAttributesAttribute", element.MCAttributes.PreserveAttributes);
-                            validationContext.EmitError(errorInfo);
-                        }
-                    }
-                }
-
-                // validate PreserveElements attribute
-                if (!string.IsNullOrEmpty(element.MCAttributes.PreserveElements))
-                {
-                    // The ProcessAttributes attribute value shall not reference any attribute name that does not belong to a namespace
-                    // that is identified by the Ignorable attribute of the same element.
-                    if (ignorableNamespaces == null)
-                    {
-                        // must have Ignorable on same element.
-                        errorInfo = validationContext.ComposeMcValidationError(element, "MC_InvalidPreserveElementsAttribute", element.MCAttributes.PreserveElements);
-                        validationContext.EmitError(errorInfo);
-                    }
-                    else
-                    {
-                        string errorQName = ValidateQNameList(element.MCAttributes.PreserveElements, ignorableNamespaces, validationContext);
-                        if (!string.IsNullOrEmpty(errorQName))
-                        {
-                            errorInfo = validationContext.ComposeMcValidationError(element, "MC_InvalidPreserveElementsAttribute", element.MCAttributes.PreserveElements);
-                            validationContext.EmitError(errorInfo);
-                        }
-                    }
-                }
-
-                // validate ProcessContent attribute
-                if (!string.IsNullOrEmpty(element.MCAttributes.ProcessContent))
-                {
-                    // The ProcessAttributes attribute value shall not reference any attribute name that does not belong to a namespace
-                    // that is identified by the Ignorable attribute of the same element.
-                    if (ignorableNamespaces == null)
-                    {
-                        // must have Ignorable on same element.
-                        errorInfo = validationContext.ComposeMcValidationError(element, "MC_InvalidProcessContentAttribute", element.MCAttributes.ProcessContent);
-                        validationContext.EmitError(errorInfo);
-                    }
-                    else
-                    {
-                        string errorQName = ValidateQNameList(element.MCAttributes.ProcessContent, ignorableNamespaces, validationContext);
-                        if (!string.IsNullOrEmpty(errorQName))
-                        {
-                            errorInfo = validationContext.ComposeMcValidationError(element, "MC_InvalidProcessContentAttribute", element.MCAttributes.ProcessContent);
-                            validationContext.EmitError(errorInfo);
-                        }
-                    }
-
-                    foreach (var exAttribute in element.ExtendedAttributes)
-                    {
-                         // Markup consumers that encounter a non-ignored element that has an xml:lang or xml:space attribute and is also identified by a ProcessContent attribute value might generate an error.
-                        if (AlternateContentValidator.IsXmlSpaceOrXmlLangAttribue(exAttribute))
-                        {
-                            // report error.
-                            errorInfo = validationContext.ComposeMcValidationError(element, "MC_InvalidXmlAttributeWithProcessContent");
-                            validationContext.EmitError(errorInfo);
-                        }
-                    }
-                }
-
-                if (!string.IsNullOrEmpty(element.MCAttributes.MustUnderstand))
-                {
-                    // TODO: MustUnderstand
-                    // A markup consumer that does not understand these identified namespaces shall not continue to process the markup document
-
-                    // rule: the prefix must already be defined.
-                    var prefixes = new ListValue<StringValue>();
-                    prefixes.InnerText = element.MCAttributes.MustUnderstand;
-                    foreach (var prefix in prefixes.Items)
-                    {
-                        var mustunderstandNamespace = element.LookupNamespace(prefix);
-                        if (string.IsNullOrEmpty(mustunderstandNamespace))
-                        {
-                            // report error, the prefix is not defined.
-                            errorInfo = validationContext.ComposeMcValidationError(element, "MC_InvalidMustUnderstandAttribute", element.MCAttributes.MustUnderstand);
-                            validationContext.EmitError(errorInfo);
-                        }
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Validate QName list in PreserveElements, PreserveAttributes, ProcessContent.
-        /// </summary>
-        /// <param name="qnameList">The QName list to be validated.</param>
-        /// <param name="ignorableNamespaces">The ignorable namespaces.</param>
-        /// <param name="validationContext"></param>
-        /// <returns>The QName that is not valid.</returns>
-        internal static string ValidateQNameList(string qnameList, HashSet<string> ignorableNamespaces, ValidationContext validationContext)
-        {
-            Debug.Assert(!string.IsNullOrEmpty(qnameList));
-
-            var qnames = new ListValue<StringValue>();
-            qnames.InnerText = qnameList;
-
-            foreach (var qname in qnames.Items)
-            {
-                // must be QName
-                var items = qname.Value.Split(':');
-                if (items.Length != 2)
-                {
-                    return qname;
-                }
-
-                // Prefix must be already defined.
-                var attributeNamesapce = validationContext.Element.LookupNamespace(items[0]);
-                if (string.IsNullOrEmpty(attributeNamesapce))
-                {
-                    return qname;
-                }
-
-                // The namespace must be identified by the Ignorable attribute at the same element.
-                if (!ignorableNamespaces.Contains(attributeNamesapce))
-                {
-                    return qname;
-                }
-            }
-
-            return null;
-        }
-    }
-
-        /// <summary>
-    /// Static class to hold extension methods for OpenXmlElement.
-    /// </summary>
-    internal static partial class OpenXmlElementExtensionMethods
-    {
-        #region MC related methods
-
-        internal static OpenXmlElement GetFirstChildMc(this OpenXmlElement parent, MCContext mcContext, FileFormatVersions format)
-        {
-            var child = parent.GetFirstNonMiscElementChild();
-            return parent.GetChildMc(child, mcContext, format);
-        }
-
-        internal static OpenXmlElement GetNextChildMc(this OpenXmlElement parent, OpenXmlElement child, MCContext mcContext, FileFormatVersions format)
-        {
-            var next = child.GetNextNonMiscElementSibling();
-            var mcTier = child.Parent;
-
-            if (next == null && mcTier != parent)
-            {
-                // the child must be under element in ProcessContent or ACB
-                if (mcTier is AlternateContentChoice || mcTier is AlternateContentFallback)
-                {
-                    mcTier = mcTier.Parent;
-                }
-                Debug.Assert(mcTier != null);
-
-                // there is no more next sibling in this level, then try to find the next siblig of the up level.
-                return parent.GetNextChildMc(mcTier, mcContext, format);
-            }
-
-            return parent.GetChildMc(next, mcContext, format);
-        }
-
-        /// <summary>
-        /// Gets the next child (skip all MC elements (skip ACB layer, skip Ignorabled element.)).
-        /// </summary>
-        /// <param name="parent">The logic parent element.</param>
-        /// <param name="child">The child element to be tested.</param>
-        /// <param name="mcContext">Markup Compatibility context.</param>
-        /// <param name="format">Targeting file format (Office2007 or Office201).</param>
-        /// <returns>The logic child (when we apply a MC pre-processer).</returns>
-        private static OpenXmlElement GetChildMc(this OpenXmlElement parent, OpenXmlElement child, MCContext mcContext, FileFormatVersions format)
-        {
-            // Use stack to cache the next siblings in different levels.
-            Stack<OpenXmlElement> nextSiblings = new Stack<OpenXmlElement>();
-
-            while (child != null)
-            {
-                var acb = child as AlternateContent;
-                if (acb == null && child.IsInVersion(format))
-                {
-                    return child;
-                }
-                else
-                {
-                    mcContext.PushMCAttributes2(child.MCAttributes, child.LookupNamespace);
-                    if (acb != null)
-                    {
-                        nextSiblings.Push(child.GetNextNonMiscElementSibling());
-                        var select = mcContext.GetContentFromACBlock(acb, format);
-                        if (select != null)
-                        {
-                            child = select.GetFirstNonMiscElementChild();
-                        }
-                        else
-                        {
-                            // The ACB has no children elements.
-                            // case like: <acb/> <acb><choice/><fallback/></acb>
-                            child = null;
-                        }
-                    }
-                    else
-                    {
-                        // Ignorable element, skip it
-                        if (mcContext.IsIgnorableNs(child.NamespaceUri))
-                        {
-                            // Any element marked with ProcessContent should be an Ignorable Element
-                            if (mcContext.IsProcessContent(child))
-                            {
-                                nextSiblings.Push(child.GetNextNonMiscElementSibling());
-                                //
-                                child = child.GetFirstNonMiscElementChild();
-                            }
-                            else
-                            {
-                                child = child.GetNextNonMiscElementSibling();
-                            }
-                        }
-                        else
-                        {
-                            mcContext.PopMCAttributes2();
-                            return child;
-                        }
-                    }
-                    mcContext.PopMCAttributes2();
-                }
-
-                while (child == null && nextSiblings.Count > 0)
-                {
-                    child = nextSiblings.Pop();
-                }
-            }
-
-            // child is null.
-            return child;
-        }
-
-        #endregion
     }
 }
