+++ DocumentFormat.OpenXml/Validation/Schema/ParticleMatch.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace DocumentFormat.OpenXml.Internal.SchemaValidation
+{
+    /// <summary>
+    /// Particle match result.
+    /// </summary>
+    internal enum ParticleMatch
+    {
+        Nomatch,
+        Partial,
+        Matched,
+    }
+}
