+++ DocumentFormat.OpenXml/Validation/Schema/SdbParticleChildrenIndex.cs
@@ -1,0 +1,85 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+/**********************************************************
+ * Define data struct for schema constraint binary database
+ **********************************************************/
+
+using System;
+using System.Diagnostics;
+
+using SdbIndex = System.UInt16;
+
+namespace DocumentFormat.OpenXml.Internal.SchemaValidation
+{
+    /// <summary>
+    /// Particle children index data.
+    /// </summary>
+    [DebuggerDisplay("ParticleIndex={ParticleIndex}")]
+    internal class SdbParticleChildrenIndex : SdbData
+    {
+        /// <summary>
+        /// The index of the particle in the SdbParticleConstraint data array.
+        /// </summary>
+        public SdbIndex ParticleIndex { get; set; }
+
+        public SdbParticleChildrenIndex()
+        {
+            this.ParticleIndex = SdbData.InvalidId;
+        }
+
+        public SdbParticleChildrenIndex(SdbIndex index)
+        {
+            this.ParticleIndex = index;
+        }
+
+        public SdbParticleChildrenIndex(int index)
+        {
+            if (index >= SdbData.MaxSdbIndex)
+            {
+                throw new ArgumentOutOfRangeException(nameof(index));
+            }
+
+            this.ParticleIndex = (SdbIndex)index;
+        }
+
+        /// <summary>
+        /// The size in bytes of this data structure.
+        /// </summary>
+        public static int TypeSize
+        {
+            get { return sizeof(SdbIndex); }
+        }
+
+        #region Override SdbData Members
+
+        /// <summary>
+        /// The size in bytes of this data structure.
+        /// </summary>
+        public override int DataSize
+        {
+            get { return TypeSize; }
+        }
+
+        /// <summary>
+        /// Serialize the data into byte data.
+        /// </summary>
+        /// <returns>Byte data.</returns>
+        public override byte[] GetBytes()
+        {
+            return this.ParticleIndex.Bytes();
+        }
+
+        /// <summary>
+        /// Deserialize the data from byte data.
+        /// </summary>
+        /// <param name="value">The byte data.</param>
+        /// <param name="startIndex">The offset the data begins at.</param>
+        public override void LoadFromBytes(byte[] value, int startIndex)
+        {
+            this.ParticleIndex = LoadSdbIndex(value, ref startIndex);
+        }
+
+        #endregion
+    }
+}
