+++ DocumentFormat.OpenXml/Validation/ValidationErrorInfo.cs
@@ -7,29 +7,6 @@ using System.Diagnostics;
 namespace DocumentFormat.OpenXml.Validation
 {
     /// <summary>
-    /// The type of the validation error.
-    /// </summary>
-    public enum ValidationErrorType
-    {
-        /// <summary>
-        /// Schema validation error.
-        /// </summary>
-        Schema,
-        /// <summary>
-        /// Semantic validation error.
-        /// </summary>
-        Semantic,
-        /// <summary>
-        /// Package structure validation error.
-        /// </summary>
-        Package,
-        /// <summary>
-        /// Markup Compatibility validation error.
-        /// </summary>
-        MarkupCompatibility,
-    }
-
-    /// <summary>
     /// Defines the ValidationErrorInfo.
     /// </summary>
     [DebuggerDisplay("Description={Description}")]
