+++ DocumentFormat.OpenXml/Validation/ValidationSettings.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace DocumentFormat.OpenXml
+namespace DocumentFormat.OpenXml.Validation
 {
     /// <summary>
     /// Settings for validation.
